Add flagging to preprocess map branch filter
